virtio_blk + kernel: RegisterGrant wrapper + APIC/IPC/exit polish
virtio_blk setup folds MMIO/PIO grant release into a single
RegisterGrant enum so each phase rolls back its own grant
without the call sites duplicating the unmap path. Adds
constants split (queue.rs), an explicit trace.rs marker module,
and tightens the bring-up sequence around the new wrapper.
Kernel-side: small fixes in x86_64 APIC xAPIC init + state,
IPC delivery, process exit-and-yield, and the virtio_blk
spawn mirror.

+++ src/arch/x86_64/interrupt/apic/init_xapic.rs
@@ -29,7 +29,7 @@ pub(super) unsafe fn init_xapic() -> ApicResult<()> {
     unsafe {
         let phys = (rdmsr(IA32_APIC_BASE) & 0xFFFF_F000) as u64;
         let va = map_apic_mmio(PhysAddr::new(phys))?;
-        MMIO_BASE.store(va.as_u64() as u32, Ordering::Release);
+        MMIO_BASE.store(va.as_u64(), Ordering::Release);
 
         mmio_w32(LAPIC_SVR, SVR_APIC_ENABLE | VEC_SPURIOUS as u32);
         mmio_w32(LAPIC_LVT_LINT0, LVT_NMI);
